requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.9
Django==4.0
django-rest-framework==0.1.0
djangorestframework==3.13.1
idna==3.3
pytz==2021.3
requests==2.26.0
soupsieve==2.3.1
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.7
